requirements.txt
streamlit==1.32.2
requests==2.31.0
PyJWT==2.8.0